testing/web-platform/tests/html/cross-origin-embedder-policy/reporting.https.html

<!doctype html>
<html>
<html>
<body>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/utils.js"></script>
<script src="/common/get-host-info.sub.js"></script>
<script>
const HOST = get_host_info();
const REMOTE_ORIGIN = HOST.REMOTE_ORIGIN;
const BASE = new URL("resources", location).pathname

function wait(ms) {
  return new Promise(resolve => step_timeout(resolve, ms));
}

async function pollReports(endpoint, reports) {
  while (true) {
    await wait(200);
    const res = await fetch(`resources/report.py?endpoint=${endpoint}`, {cache: 'no-store'});
    if (res.status !== 200) {
      continue;
    }
    for (const report of await res.json()) {
      reports.push(report);
    }
  }
}

const reports = [];
const reportsForReportOnly = [];
pollReports('endpoint', reports);
pollReports('report-only-endpoint', reportsForReportOnly);

function checkCorpReportExistence(reports, blockedUrl, contextUrl) {
  blockedUrl = new URL(blockedUrl, location).href;
  contextUrl = new URL(contextUrl, location).href;
  for (const report of reports) {
    if (report.type !== 'coep' || report.url !== contextUrl ||
        report.body.type !== 'corp') {
      continue;
    }
    if (report.body['blocked-url'] === blockedUrl) {
      return;
    }
  }
  assert_unreached(`A report whose blocked-url is ${blockedUrl} and url is ${contextUrl} is not found.`);
}

function checkNavigationReportExistence(reports, blockedUrl, contextUrl) {
  blockedUrl = new URL(blockedUrl, location).href;
  contextUrl = new URL(contextUrl, location).href;
  for (const report of reports) {
    if (report.type !== 'coep' || report.url !== contextUrl ||
        report.body.type !== 'navigation') {
      continue;
    }
    if (report.body['blocked-url'] === blockedUrl) {
      return;
    }
  }
  assert_unreached(`A report whose blocked-url is ${blockedUrl} and url is ${contextUrl} is not found.`);
}

function checkReportNonExistence(reports, blockedUrl, contextUrl) {
  blockedUrl = new URL(blockedUrl, location).href;
  contextUrl = new URL(contextUrl, location).href;
  for (const report of reports) {
    if (report.type !== 'coep' || report.url !== contextUrl) {
      continue;
    }
    assert_not_equals(report.body['blocked-url'], blockedUrl);
  }
}

async_test(async (t) => {
  try {
    const iframe = document.createElement('iframe');
    t.add_cleanup(() => iframe.remove());

    iframe.src = `resources/reporting-empty-frame.html`
    document.body.appendChild(iframe);
    await new Promise(resolve => {
      iframe.addEventListener('load', resolve, {once: true});
    });

    function fetchInIframe(url) {
      const init = { mode: 'no-cors', cache: 'no-store' };
      iframe.contentWindow.fetch(url, init).catch(() => {});
    }

    const suffix = 'subresource-corp';
    const sameOriginUrl = `/common/text-plain.txt?${suffix}`;
    const blockedByPureCorp = `${REMOTE_ORIGIN}${BASE}/nothing-same-origin-corp.txt?${suffix}`;
    const blockedDueToCoep = `${REMOTE_ORIGIN}/common/text-plain.txt?abc&${suffix}`;
    const dest = `${REMOTE_ORIGIN}/common/text-plain.txt?xyz&${suffix}`;
    const redirect = `/common/redirect.py?location=${encodeURIComponent(dest)}&${suffix}`;

    fetchInIframe(sameOriginUrl);
    fetchInIframe(blockedByPureCorp);
    fetchInIframe(blockedDueToCoep);
    fetchInIframe(redirect);

    // Wait 3 seconds for reports to settle.
    await wait(3000);

    checkReportNonExistence(reports, sameOriginUrl, iframe.src);
    checkReportNonExistence(reports, blockedByPureCorp, iframe.src);
    checkCorpReportExistence(reports, blockedDueToCoep, iframe.src);
    checkCorpReportExistence(reports, redirect, iframe.src);
    checkReportNonExistence(reports, dest, iframe.src);

    t.done();
  } catch (e) {
    t.step(() => { throw e });
  }
}, 'subresource CORP');

async_test(async (t) => {
  try {
    const iframe = document.createElement('iframe');
    t.add_cleanup(() => iframe.remove());

    iframe.src = `resources/reporting-empty-frame.html`
    document.body.appendChild(iframe);
    await new Promise(resolve => {
      iframe.addEventListener('load', resolve, {once: true});
    });

    const w = iframe.contentWindow;

    function attachFrame(url) {
      const frame = w.document.createElement('iframe');
      frame.src = url;
      w.document.body.appendChild(frame);
    }

    const suffix = 'navigation-corp';
    const coep = `pipe=header(cross-origin-embedder-policy,require-corp)`;
    const sameOrigin = `/common/blank.html?${coep}&${suffix}`;
    const blockedDueToCoep = `${REMOTE_ORIGIN}/common/blank.html?${coep}&${suffix}-a`;
    const dest = `${REMOTE_ORIGIN}/common/blank.html?${coep}&${suffix}-b`;
    const redirect = `/common/redirect.py?location=${encodeURIComponent(dest)}&${suffix}`;

    attachFrame(sameOrigin);
    attachFrame(blockedDueToCoep);
    attachFrame(redirect);

    // Wait 3 seconds for reports to settle.
    await wait(3000);

    checkReportNonExistence(reports, sameOrigin, iframe.src);
    checkCorpReportExistence(reports, blockedDueToCoep, iframe.src);
    checkCorpReportExistence(reports, redirect, iframe.src);
    checkReportNonExistence(reports, dest, iframe.src);

    t.done();
  } catch (e) {
    t.step(() => { throw e });
  }
}, 'navigation CORP');

async_test(async (t) => {
  try {
    const iframe = document.createElement('iframe');
    t.add_cleanup(() => iframe.remove());

    const suffix = '&navigation-coep';
    const corp = 'header(cross-origin-resource-policy,cross-origin)';
    const noCoep = `pipe=${corp}`;
    const coep =
      `pipe=header(cross-origin-embedder-policy,require-corp%3breport-to=%22endpoint%22)|${corp}`;
    const coepReportOnly =
      `pipe=header(cross-origin-embedder-policy-report-only,require-corp%3breport-to=%22report-only-endpoint%22)|${corp}`;
    const path = `/common/blank.html`;
    const pipes = [noCoep, coep, coepReportOnly];
    const settings = new Map();
    settings.set(noCoep, {
      pipe: noCoep,
      value: 'unsafe-none',
      reportOnlyValue: 'unsafe-none',
    });
    settings.set(coep, {
      pipe: coep,
      value: 'require-corp',
      reportOnlyValue: 'unsafe-none',
    });
    settings.set(coepReportOnly, {
      pipe: coepReportOnly,
      value: 'unsafe-none',
      reportOnlyValue: 'require-corp',
    });

    function genUrl(pipe) {
      return `${path}?${pipe}${suffix}`;
    }

    for (const outer of settings.keys()) {
      for (const inner of settings.keys()) {
        const iframe = document.createElement('iframe');
        t.add_cleanup(() => iframe.remove());

        iframe.src = genUrl(outer);
        iframe.addEventListener('load', () => {
          const w = iframe.contentWindow;
          const d = iframe.contentDocument;
          const nested = d.createElement('iframe');
          nested.src = genUrl(inner) + '-nested';
          d.body.appendChild(nested);
        }, {once: true});
        document.body.appendChild(iframe);
      }
    }

    // Wait 3 seconds for reports to settle.
    await wait(3000);

    function check(rs, inner, outer) {
      checkNavigationReportExistence(
        rs, genUrl(inner) + '-nested', genUrl(outer));
    }
    function checkNoReport(reports, inner, outer) {
      checkReportNonExistence(
        reports, genUrl(inner) + '-nested', genUrl(outer));
    }

    // outer === noCoep
    checkNoReport(reports, noCoep, noCoep);
    checkNoReport(reports, coep, noCoep);
    checkNoReport(reports, coepReportOnly, noCoep);
    checkNoReport(reportsForReportOnly, noCoep, noCoep);
    checkNoReport(reportsForReportOnly, coep, noCoep);
    checkNoReport(reportsForReportOnly, coepReportOnly, noCoep);

    // outer === coep
    check(reports, noCoep, coep);
    checkNoReport(reports, coep, coep);
    check(reports, coepReportOnly, coep);
    checkNoReport(reportsForReportOnly, noCoep, coep);
    checkNoReport(reportsForReportOnly, coep, coep);
    checkNoReport(reportsForReportOnly, coepReportOnly, coep);

    // outer === coepReportOnly
    checkNoReport(reports, noCoep, coepReportOnly);
    checkNoReport(reports, coep, coepReportOnly);
    checkNoReport(reports, coepReportOnly, coepReportOnly);
    check(reportsForReportOnly, noCoep, coepReportOnly);
    checkNoReport(reportsForReportOnly, coep, coepReportOnly);
    check(reportsForReportOnly, coepReportOnly, coepReportOnly);

    t.done();
  } catch (e) {
    t.step(() => { throw e });
  }
}, 'COEP violation on nested frame navigation');

</script>$
